BREAKING: Move useIsMounted to component-helpers
