+++ lib/component-helpers/package.json
@@ -18,7 +18,11 @@
   },
   "author": "Daniel A.C. Martin <npm@daniel-martin.co.uk> (http://daniel-martin.co.uk/)",
   "license": "MIT",
+  "peerDependencies": {
+    "react": "^16.14.0 || ^17.0.0 || ^18.0.0"
+  },
   "devDependencies": {
+    "@types/react": "^18.3.11",
     "typescript": "4.9.5"
   }
 }
