+++ lib/route-utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './active';
-export * from './is-mounted';
 export { queryString } from './query-string';
 export * from './router';
 export * from './url-parse';
